menata ulang namespace dan membuat auto connection
