+++ App.config
@@ -1,8 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <configuration>
 	<connectionStrings>
-		<add name ="ConnString" 
+		<add name ="ZhafarConnString" 
 			 connectionString="Data Source=MSI\SQLEXPRESS;Initial Catalog=db_hotel_terakhir;Integrated Security=True" 
+			 providerName="System.Data.SqlClient"/>
+		<add name ="PijarConnString" 
+			 connectionString="Data Source=.\sqlexpress;Initial Catalog=db_hotel;Integrated Security=True" 
 			 providerName="System.Data.SqlClient"/>
 		<!-- Add your own connection string here -->
