+++ DataAccessLayer/RoomFacilityBathroomDAL.cs
@@ -3,23 +3,33 @@ using System;
 using System.Collections.Generic;
 using System.Configuration;
 using System.Data.SqlClient;
-using System.Diagnostics;
 using System.Linq;
 
-namespace PinusPengger.Model.DataAccessLayer
+namespace PinusPengger.DataAccessLayer
 {
     /// <summary>
-    /// Mekanisme CRUD tabel kamar
+    /// Mekanisme CRUD tabel fasilitas kamar mandi
     /// </summary>
-    public class RoomDAL : IRepository
+    public class RoomFacilityBathroomDAL : IRepository
     {
-        /// <summary>
-        /// Menginisialisasi objek <see cref="RoomDAL"/>
-        /// </summary>
-        public RoomDAL()
+        private void Connect(ConnectionStringSettingsCollection settingsCollection, int index)
         {
-            Connection = new SqlConnection(ConfigurationManager.ConnectionStrings["ConnString"].ToString());
-            Connection.Open();
+            try
+            {
+                Connection = new SqlConnection(settingsCollection[index].ConnectionString);
+                Connection.Open();
+            }
+            catch (Exception)
+            {
+                Connect(settingsCollection, index + 1);
+            }
+        }
+        /// <summary>
+        /// Menginisialisasi objek <see cref="RoomFacilityBathroomDAL"/>
+        /// </summary>
+        public RoomFacilityBathroomDAL()
+        {
+            Connect(ConfigurationManager.ConnectionStrings, 0);
         }
 
         /// <summary>
@@ -33,9 +43,9 @@ namespace PinusPengger.Model.DataAccessLayer
         /// <returns><inheritdoc/></returns>
         public List<object> ReadData()
         {
-            var result = new List<Room>();
+            var result = new List<RoomFacilityBathroom>();
 
-            string query = ConfigurationManager.AppSettings["RoomDAL:ReadData"] ?? string.Empty;
+            string query = ConfigurationManager.AppSettings["RoomFacilityBathroomDAL:ReadData"] ?? string.Empty;
 
             using (var cmd = new SqlCommand(query, Connection))
             {
@@ -43,31 +53,28 @@ namespace PinusPengger.Model.DataAccessLayer
                 {
                     while (rdr.Read())
                     {
-                        var room = new Room
+                        var roomFacilityBathroom = new RoomFacilityBathroom()
                         {
-                            RoomID = rdr.GetInt32(0),
-                            RoomCode = rdr.GetString(1),
-                            RoomFloor = rdr.GetByte(2),
-                            RoomNumber = rdr.GetByte(3),
-                            SquareMeter = rdr.GetByte(4),
-                            RoomType = (Tag.RoomType)Enum.Parse(typeof(Tag.RoomType), rdr.GetString(5))
+                            NameOfFacility = rdr.GetString(0),
+                            RoomType = (Tag.RoomType)Enum.Parse(typeof(Tag.RoomType), rdr.GetString(1))
                         };
-                        result.Add(room);
+                        result.Add(roomFacilityBathroom);
                     }
                 }
             }
-            Debug.WriteLine($"RoomDAL Count: {result.Count}");
+
             return result.Select(x => (object)x).ToList();
         }
         /// <summary>
         /// <inheritdoc/>
         /// </summary>
-        /// <param name="obj"><inheritdoc/></param>
+        /// <param name="obj"></param>
         /// <exception cref="NotImplementedException"></exception>
         public void InsertRecord(object obj)
         {
             throw new NotImplementedException();
         }
+
         /// <summary>
         /// <inheritdoc/>
         /// </summary>
@@ -77,6 +84,7 @@ namespace PinusPengger.Model.DataAccessLayer
         {
             throw new NotImplementedException();
         }
+
         /// <summary>
         /// <inheritdoc/>
         /// </summary>
