+++ DataAccessLayer/RoomFacilityOtherDAL.cs
@@ -5,20 +5,31 @@ using System.Configuration;
 using System.Data.SqlClient;
 using System.Linq;
 
-namespace PinusPengger.Model.DataAccessLayer
+namespace PinusPengger.DataAccessLayer
 {
     /// <summary>
-    /// Mekanisme CRUD tabel fasilitas kamar
+    /// Mekanisme CRUD tabel fasilitas lainnya
     /// </summary>
-    public class RoomFacilityDAL : IRepository
+    public class RoomFacilityOtherDAL : IRepository
     {
-        /// <summary>
-        /// Menginisiasi objek <see cref="RoomFacilityDAL"/>
-        /// </summary>
-        public RoomFacilityDAL()
+        private void Connect(ConnectionStringSettingsCollection settingsCollection, int index)
         {
-            Connection = new SqlConnection(ConfigurationManager.ConnectionStrings["ConnString"].ToString());
-            Connection.Open();
+            try
+            {
+                Connection = new SqlConnection(settingsCollection[index].ConnectionString);
+                Connection.Open();
+            }
+            catch (Exception)
+            {
+                Connect(settingsCollection, index + 1);
+            }
+        }
+        /// <summary>
+        /// Menginisialisasi objek <see cref="RoomFacilityOtherDAL"/>
+        /// </summary>
+        public RoomFacilityOtherDAL()
+        {
+            Connect(ConfigurationManager.ConnectionStrings, 0);
         }
 
         /// <summary>
@@ -32,9 +43,9 @@ namespace PinusPengger.Model.DataAccessLayer
         /// <returns><inheritdoc/></returns>
         public List<object> ReadData()
         {
-            var result = new List<RoomFacility>();
+            var result = new List<RoomFacilityOther>();
 
-            string query = ConfigurationManager.AppSettings["RoomFacilityDAL:ReadData"] ?? string.Empty;
+            string query = ConfigurationManager.AppSettings["RoomFacilityOtherDAL:ReadData"] ?? string.Empty;
 
             using (var cmd = new SqlCommand(query, Connection))
             {
@@ -42,20 +53,18 @@ namespace PinusPengger.Model.DataAccessLayer
                 {
                     while (rdr.Read())
                     {
-                        var roomFacility = new RoomFacility()
+                        var roomFacilityOther = new RoomFacilityOther()
                         {
-                            RoomType = (Tag.RoomType)Enum.Parse(typeof(Tag.RoomType), rdr.GetString(0)),
-                            Bed = rdr.GetString(1),
-                            Internet = rdr.GetString(2)
+                            NameOfFacility = rdr.GetString(0),
+                            RoomType = (Tag.RoomType)Enum.Parse(typeof(Tag.RoomType), rdr.GetString(1))
                         };
-                        result.Add(roomFacility);
+                        result.Add(roomFacilityOther);
                     }
                 }
             }
 
             return result.Select(x => (object)x).ToList();
         }
-
         /// <summary>
         /// <inheritdoc/>
         /// </summary>
@@ -65,7 +74,6 @@ namespace PinusPengger.Model.DataAccessLayer
         {
             throw new NotImplementedException();
         }
-
         /// <summary>
         /// <inheritdoc/>
         /// </summary>
@@ -75,7 +83,6 @@ namespace PinusPengger.Model.DataAccessLayer
         {
             throw new NotImplementedException();
         }
-
         /// <summary>
         /// <inheritdoc/>
         /// </summary>
@@ -85,7 +92,6 @@ namespace PinusPengger.Model.DataAccessLayer
         {
             throw new NotImplementedException();
         }
-
         /// <summary>
         /// <inheritdoc/>
         /// </summary>
