+++ View/Pages/HistoryPage.xaml
@@ -1,17 +1,17 @@
-﻿<Page x:Class="PinusPengger.View.HomePage"
+﻿<Page x:Class="PinusPengger.View.Pages.HistoryPage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:PinusPengger.View.Pages"
       xmlns:viewModel="clr-namespace:PinusPengger.ViewModel.BasePageVM"
-      xmlns:model="clr-namespace:PinusPengger.Model"
       xmlns:icon="http://metro.mahapps.com/winfx/xaml/iconpacks"
-      xmlns:sys="clr-namespace:System;assembly=mscorlib"
       mc:Ignorable="d" 
-      d:DesignHeight="420" 
-      d:DesignWidth="920">
+      d:DesignHeight="420" d:DesignWidth="920"
+      Title="HistoryPage">
+
     <Page.DataContext>
-        <viewModel:DisplayReservationsVM/>
+        <viewModel:DisplayHistoriesVM/>
     </Page.DataContext>
     <Grid Background="White">
         <Grid.RowDefinitions>
@@ -57,28 +57,29 @@
                           Height="18"
                           HorizontalAlignment="Center"
                           VerticalAlignment="Center"
-                          Margin="0,0,0,1"/>
+                          Margin="0,0,0,1" />
                 </Button>
             </StackPanel>
         </StackPanel>
-        <DataGrid Margin="10"
+        <DataGrid x:Name="tabelHistori"
+                  Margin="10"
                   Height="290"
                   AutoGenerateColumns="False"
                   Grid.Row="3"
                   CanUserAddRows="False"
                   IsReadOnly="True"
-                  ItemsSource="{Binding ReservationJoinedsObservable}">
+                  ItemsSource="{Binding HistoryJoinedsObservable}">
             <DataGrid.Columns>
                 <DataGridTextColumn Header="Kode Pemesanan" 
-                                    Binding="{Binding Reservation.ReservationCode}"/>
+                                    Binding="{Binding History.ReservationCode}"/>
                 <DataGridTextColumn Header="Nama" Width="200" 
                                     Binding="{Binding Customer.FullName}"/>
                 <DataGridTextColumn Header="Nomor Kamar" Width="*"  
                                     Binding="{Binding Room.RoomCode}"/>
                 <DataGridTextColumn Header="Tanggal Checkin" Width="*" 
-                                    Binding="{Binding Reservation.Checkin, StringFormat='dd/MM/yyyy'}"/>
+                                    Binding="{Binding History.Checkin, StringFormat='dd/MM/yyyy'}"/>
                 <DataGridTextColumn Header="Tanggal Checkout" Width="*"  
-                                    Binding="{Binding Reservation.Checkout, StringFormat='dd/MM/yyyy'}"/>
+                                    Binding="{Binding History.Checkout, StringFormat='dd/MM/yyyy'}"/>
             </DataGrid.Columns>
         </DataGrid>
     </Grid>
